Fix create air current (#1348)
* extend create air current maxDistance to avoid randomness burn on the end of the current

* fix fabric remap
